Fix: Address Windows build errors and compiler warnings
This commit resolves several build failures and compiler warnings encountered on the Windows platform, primarily related to DLL export/import, PIMPL idiom usage, and CMake test configuration.

Key changes include:
- Ensured proper compilation and linking of non-inline Windows-specific utility functions (`win32_to_long_path`, `win32_make_unique_suffix`).
- Corrected argument passing for JsonConfig's `with_json_write` in tests.
- Implemented robust PIMPL handling for `JsonConfig` and its nested `ReadLock`/`WriteLock` classes by explicitly defining special member functions and applying `PYLABHUB_UTILS_EXPORT` macros, resolving "unresolved external symbol" linker errors.
- Adjusted CMake's test discovery working directory to ensure test executables find required DLLs at runtime on Windows, fixing "DLL Not Found" errors.
- Systematically addressed compiler warnings:
    - Suppressed C4251 (PIMPL false positive) in `JsonConfig.hpp`.
    - Corrected C4267 (size_t to int conversion) in `Lifecycle.cpp` by using `size_t`.
    - Removed unused C4189 variable in `test_process_utils.cpp`.
    - Suppressed C4996 (unsafe getenv) locally in `shared_test_helpers.cpp` for test-only code.
    - Mitigated C4172 (false positive return) in `JsonConfig.cpp`.

+++ cpp/src/utils/Lifecycle.cpp
@@ -158,7 +158,7 @@ class LifecycleManagerImpl
         std::function<void()> startup;
         InternalModuleShutdownDef shutdown;
 
-        int in_degree = 0;
+        size_t in_degree = 0;
         std::vector<InternalGraphNode *> dependents;
     };
